Restore and reorganize test files in monorepo structure
- Move calendar-specific tests to appropriate pack directories
- Fantasy pack tests: dark-sun, exandrian, forbidden-lands, warhammer, WFRP
- Sci-fi pack tests: star-trek calendar comprehensive test suite
- Update vitest config to include all package test directories
- Update bundle analysis workflow for monorepo structure
- Restore missing tests: 836 → 931 tests (95 tests recovered)

All test files updated with correct import paths for monorepo structure.
Bundle analysis now covers core module + both calendar packs separately.
