+++ packages/fantasy-pack/test/forbidden-lands.test.ts
@@ -7,14 +7,14 @@
  */
 
 import { describe, test, expect, beforeEach } from 'vitest';
-import { CalendarEngine } from '../../src/core/calendar-engine';
-import type { SeasonsStarsCalendar } from '../../src/types/calendar';
+import { CalendarEngine } from '../../core/src/core/calendar-engine';
+import type { SeasonsStarsCalendar } from '../../core/src/types/calendar';
 import * as fs from 'fs';
 import * as path from 'path';
 
 // Load Forbidden Lands calendar for testing
 function loadCalendar(): SeasonsStarsCalendar {
-  const calendarPath = path.join('packages/fantasy-pack/calendars', 'forbidden-lands.json');
+  const calendarPath = path.join(__dirname, '../calendars', 'forbidden-lands.json');
   const calendarData = JSON.parse(fs.readFileSync(calendarPath, 'utf8'));
   return calendarData;
 }
